+++ Kalimbra/Instruments/IInstrument.cs
@@ -2,6 +2,6 @@ namespace Kalimbra.Instruments
 {
     public interface IInstrument
     {
-        double[] Play(Note note);
+        int[] Play(Note note);
     }
 }
